Fixed circular dependency build issue
